+++ server/src/Birdism/Cache.hs
@@ -15,18 +15,18 @@ module Birdism.Cache
   ) where
 
 import           Birdism.Common      (MonadIO, Text, liftIO, (<&>))
+import           Birdism.Types       (Family, Region, SpeciesCode)
 import           Control.Lens        (makeClassy)
-import Birdism.Types (Region, Family)
 
 import qualified Data.HashMap.Strict as Map
 import qualified Data.IORef          as IORef
 
 data CacheData
   = CacheData
-  { cdRegionsCache :: [Region]
+  { cdRegionsCache  :: [Region]
   , cdFamiliesCache :: [Family]
-  , cdFlickrCache :: Map.HashMap Text [Text]
-  , cdEbirdCache :: Map.HashMap Text [Text]
+  , cdFlickrCache   :: Map.HashMap Text [Text]
+  , cdEbirdCache    :: Map.HashMap Text [SpeciesCode]
   }
 
 newtype BirdismCache =
@@ -65,12 +65,12 @@ writeFlickrCache key val store =
   liftIO $ IORef.modifyIORef' (unBirdisimCache store) $ \cache ->
     cache { cdFlickrCache = Map.insert key val $ cdFlickrCache cache }
 
-lookupEbirdCache :: MonadIO m => Text -> BirdismCache -> m (Maybe [Text])
+lookupEbirdCache :: MonadIO m => Text -> BirdismCache -> m (Maybe [SpeciesCode])
 lookupEbirdCache key store = do
   cache <- lookupp store <&> cdEbirdCache
   pure $ Map.lookup key cache
 
-writeEbirdCache :: MonadIO m => Text -> [Text] -> BirdismCache -> m ()
+writeEbirdCache :: MonadIO m => Text -> [SpeciesCode] -> BirdismCache -> m ()
 writeEbirdCache key val store =
   liftIO $ IORef.modifyIORef' (unBirdisimCache store) $ \cache ->
     cache { cdEbirdCache = Map.insert key val $ cdEbirdCache cache }