+++ server/src/Service/Ebird/V2/Product.hs
@@ -3,12 +3,15 @@ module Service.Ebird.V2.Product
   ( getSpeciesListByRegion
   ) where
 
+import qualified Birdism.Cache           as Cache
 import qualified Data.Text               as T
 
 import           Birdism.Common
 import           Birdism.Config
 import           Birdism.Types
+import           Control.Lens            ((^.))
 import           Service.Ebird.V2.Common
+
 
 speciesListByRegionUrl :: Text -> String
 speciesListByRegionUrl region = "https://api.ebird.org/v2/product/spplist/" <> T.unpack region
@@ -17,9 +20,18 @@ speciesListByRegionUrl region = "https://api.ebird.org/v2/product/spplist/" <> T
 getSpeciesListByRegion
   :: ( MonadReader r m
      , HasEBirdConf r
+     , HasBirdismCache r
      , MonadError e m
      , AsEbirdError e
      , MonadIO m
      )
   => RegionCode -> m [SpeciesCode]
-getSpeciesListByRegion = ebirdApiGetService . speciesListByRegionUrl . uRegionCode
+getSpeciesListByRegion region = do
+  cache <- asks (^. birdismCache)
+  let reg = uRegionCode region
+  Cache.lookupEbirdCache reg cache >>= \case
+    Just spList -> pure spList
+    Nothing -> do
+      spList <- ebirdApiGetService $ speciesListByRegionUrl reg
+      Cache.writeEbirdCache reg spList cache
+      pure spList